Fixed the code to 16-bit sequence numbers

+++ testing/testmedia.hpp
@@ -75,7 +75,7 @@ protected:
 
     struct ReadPos
     {
-        int32_t sequence;
+        uint16_t sequence;
         bytevector packet;
     };
     map<SRTSOCKET, ReadPos> m_group_positions;
